Rename decorator files
