+++ test/unit/decorators/test_layer_settings.py
@@ -4,7 +4,7 @@ from layer.contracts.assets import AssetType
 from layer.contracts.fabrics import Fabric
 from layer.decorators import dataset, fabric, pip_requirements
 from layer.exceptions.exceptions import ConfigError
-from layer.settings import LayerSettings
+from layer.decorators.settings import LayerSettings
 
 
 settings = LayerSettings()
